backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
beautifulsoup4
requests
selenium
aiofiles
python-multipart
sqlalchemy
